+++ Lesson8HW/src/main/java/ru/otus/lesson8hw/services/BookService.java
@@ -8,11 +8,11 @@ import java.util.List;
  * Created by zhmv on 01.08.2018.
  */
 public interface BookService {
-    int count();
-    Book getById(int id);
+    long count();
+    Book getById(long id);
     List<Book> getAllBook();
     void insert(String name, String description,
-                int idAuthor, String firstName, String secondName,
-                int idGenre, String genreName);
-    void deleteById(int id);
+                long idAuthor, String firstName, String secondName,
+                long idGenre, String genreName);
+    void deleteById(long id);
 }
