+++ Lesson8HW/src/main/java/ru/otus/lesson8hw/services/GenreService.java
@@ -9,9 +9,9 @@ import java.util.List;
  */
 public interface GenreService {
 
-    int count();
-    Genre getById(int id);
+    long count();
+    Genre getById(long id);
     List<Genre> getAllGenre();
     void insert(String genreName);
-    void deleteById(int id);
+    void deleteById(long id);
 }
